+++ tests/express.spec.ts
@@ -90,6 +90,7 @@ describe('Connect-style Requests (Express)', () => {
                 perPage: 10,
                 currentPage: 1,
                 lastPage: 10,
+                path: '/users',
             },
         }
 
@@ -101,13 +102,14 @@ describe('Connect-style Requests (Express)', () => {
         expect(response.body).toEqual({
             data: resource.data,
             links: {
-                last: 10,
+                last: 'https://localhost/users?page=10',
             },
             meta: {
                 total: 100,
                 per_page: 10,
                 current_page: 1,
                 last_page: 10,
+                path: '/users',
             },
         })
     })
@@ -131,6 +133,7 @@ describe('Connect-style Requests (Express)', () => {
                 perPage: 10,
                 currentPage: 1,
                 lastPage: 1,
+                path: '/users',
             },
         }
 
@@ -142,13 +145,14 @@ describe('Connect-style Requests (Express)', () => {
         expect(response.body).toEqual({
             data: resource.data,
             links: {
-                last: 1,
+                last: 'https://localhost/users?page=1',
             },
             meta: {
                 total: 0,
                 per_page: 10,
                 current_page: 1,
                 last_page: 1,
+                path: '/users',
             },
         })
     })